+++ fetcher/database/env.py
@@ -10,6 +10,7 @@ from sqlalchemy import engine_from_config, pool
 
 # local
 from fetcher.config import conf, fix_database_url
+from fetcher.database import models
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -24,9 +25,7 @@ fileConfig(config.config_file_name)
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-# from myapp import mymodel
-# target_metadata = mymodel.Base.metadata
-target_metadata = None
+target_metadata = models.Base.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired: